requirements.txt
openai==1.95.1
anthropic==0.57.1
python-dotenv==1.1.1
pydantic==2.11.7
tenacity==9.1.2
structlog==25.4.0
pytest==8.4.1
pytest-asyncio==1.0.0
httpx==0.28.1